+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/color-picker/color-picker.tpl
