+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/column-list-view.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb.js');
-var CustomListView = require('../../../../custom-list/custom-view');
-var CustomListCollection = require('../../../../custom-list/custom-list-collection');
+var CustomListView = require('../../../custom-list/custom-view');
+var CustomListCollection = require('../../../custom-list/custom-list-collection');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
