--- lib/assets/javascripts/cartodb3/components/form-components/editors/fill/fill-tab.tpl
@@ -1,1 +1,0 @@
-Adipisicing exercitationem sequi at fuga tempora. Voluptates numquam ex nostrum asperiores culpa. Molestiae debitis impedit corrupti amet quibusdam officia praesentium! Itaque ipsum cum aliquam doloremque laborum. Accusamus cupiditate ipsa quam.
