+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-color-dialog-content.js
@@ -2,10 +2,10 @@ var _ = require('underscore');
 var cdb = require('cartodb.js');
 var tabPaneTemplate = require('../fill-tab-pane.tpl');
 var createTextLabelsTabPane = require('../../../../../components/tab-pane/create-text-labels-tab-pane');
-var ColorPickerView = require('./color-picker');
+var ColorPickerView = require('../color-picker/color-picker');
 var InputColorValueContentView = require('./input-color-value-content-view');
 
-var rampList = require('./ramps');
+var rampList = require('../input-ramp/ramps');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
