+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-color-list.tpl
