+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-color/input-color-value-content-view.js
@@ -1,9 +1,8 @@
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
 var StackLayoutView = require('../../../../../components/stack-layout/stack-layout-view');
-var ColumnListView = require('./column-list-view');
-var InputRampContentView = require('./input-ramp-content-view');
-// var rampList = require('./ramps');
+var ColumnListView = require('../column-list-view');
+var InputRampContentView = require('../input-ramp/input-ramp-content-view');
 
 var QUANTIFICATION_METHODS = ['Jenks', 'Equal Interval', 'Heads/Tails', 'Quantile']; // TODO: translate
 var BINS = ['3', '4', '5', '6', '7']; // TODO: translate
