+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-number/input-number-value-content-view.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
 var StackLayoutView = require('../../../../../components/stack-layout/stack-layout-view');
-var ColumnListView = require('./column-list-view');
+var ColumnListView = require('../column-list-view');
 var InputNumberContentView = require('./input-number-content-view');
 
 var QUANTIFICATION_METHODS = ['Jenks', 'Equal Interval', 'Heads/Tails', 'Quantile']; // TODO: translate
