+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-number/input-number.js
