+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-ramp/input-ramp-content-view.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb.js');
 var template = require('./input-ramp-content-view.tpl');
-var RampListView = require('./ramp-list-view');
+var RampListView = require('./input-ramp-list-view');
 
 module.exports = cdb.core.View.extend({
   events: {
