+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-ramp/input-ramp-list-item-template.tpl
