+++ lib/assets/javascripts/cartodb3/components/form-components/editors/fill/input-ramp/input-ramp-list-view.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var cdb = require('cartodb.js');
 var CustomListView = require('../../../../custom-list/custom-view');
 var CustomListCollection = require('../../../../custom-list/custom-list-collection');
-var rampItemTemplate = require('./ramp-list-item-template.tpl');
+var rampItemTemplate = require('./input-ramp-list-item-template.tpl');
 var rampList = require('./ramps');
 
 module.exports = cdb.core.View.extend({
