+++ lib/assets/test/spec/cartodb3/components/form-components/editors/fill.spec.js
@@ -5,7 +5,7 @@ Backbone.$ = $;
 require('../../../../../../javascripts/cartodb3/components/form-components/editors/base.js');
 require('../../../../../../javascripts/cartodb3/components/form-components/editors/fill.js');
 
-describe('components/form-components/editors/fill', function () {
+fdescribe('components/form-components/editors/fill', function () {
   beforeEach(function () {
     this.model = new Backbone.Model({
       names: 'pepe'
@@ -14,7 +14,22 @@ describe('components/form-components/editors/fill', function () {
     this.view = new Backbone.Form.editors.Fill({
       key: 'names',
       schema: {
-        columns: ['pepe', 'paco', 'juan']
+        columns: ['pepe', 'paco', 'juan'],
+        inputs: {
+          size: {
+            range: [1, 30],
+            attribute: 'the_geom',
+            quantification: 'Quantile'
+          },
+          color: {
+            bins: 5,
+            range: ['#FFF', '#FABADA', '#00FF00', '#000', '#99999'],
+            attribute: 'name',
+            operation: 'multiply',
+            quantification: 'Jenks',
+            opacity: 0.5
+          }
+        }
       },
       model: this.model
     });
@@ -23,6 +38,8 @@ describe('components/form-components/editors/fill', function () {
 
   it('should render properly', function () {
     expect(this.view.$('.CDB-OptionInput-item').length).toBe(2);
+    expect(this.view.$el.text()).toContain('1..30');
+    expect(this.view.$el).toContain('#00FF00');
   });
 
   afterEach(function () {
